requirements.txt
openwisp-controller~=0.8
django-model-utils<4.1.0,>=4.0.0